fix checks
