+++ packages/web-ui/src/RadioGroup/RadioItem.tsx
@@ -1,13 +1,71 @@
 import { forwardRef, useContext } from 'react';
-import type { RefObject } from 'react';
-import { Box } from '../Box';
+import type { RefObject, ReactNode } from 'react';
+import { Box, BoxProps } from '../Box';
 import { useFocusRing, useRadio, useLabel, useId } from 'react-aria';
 import { RadioGroupContext } from './RadioGroup';
 import { FieldLabel } from '../FieldLabel';
 import { FormHelperText } from '../FormHelperText';
 import { colors, colorsCommon } from '@utilitywarehouse/colour-system';
 import { transition } from '../tokens';
-import { RadioInput, RadioItemProps } from './Radio';
+import type { AriaRadioProps } from 'react-aria';
+import { styled } from '../theme';
+
+/**
+ * RadioInput is an internal component used in the RadioItem & RadioTile
+ * components. It is not intended for public use.
+ *
+ * The RadioInput renders the HTML `input` element, but visually hiding it.
+ * https://moderncss.dev/pure-css-custom-styled-radio-buttons/
+ *
+ * It then renders the internal filled circle shown when the radio is checked,
+ * as well as the larger containing circle to support hover & focus styles.
+ */
+// export const RadioInput = props => <Box component="input" />;
+
+export const RadioInput: any = styled('input')(() => {
+  return {
+    // visually hidden styles
+    appearance: 'none',
+    backgroundColor: 'transparent',
+    margin: 0,
+    // radio styles
+    cursor: 'pointer',
+    outline: 'none',
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    zIndex: 1,
+    width: '100%',
+    height: '100%',
+    borderRadius: '50%',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    color: colors.cyan400,
+    '&:before': {
+      content: '""',
+      width: 14,
+      height: 14,
+      borderRadius: '50%',
+      transform: 'scale(0)',
+      transition: 'transform 200ms cubic-bezier(0.4, 0, 0.2, 1) 0ms',
+      boxShadow: `inset 1em 1em currentColor`,
+    },
+    '&:checked:before': {
+      transform: 'scale(1)',
+    },
+    '&:disabled': {
+      cursor: 'auto',
+      color: colors.grey300,
+    },
+  };
+});
+
+export interface RadioItemProps extends Omit<AriaRadioProps, 'isDisabled'> {
+  disabled?: AriaRadioProps['isDisabled'];
+  sx?: BoxProps['sx'];
+  helperText?: ReactNode;
+}
 
 /**
  * The `RadioItem` should be used within a `RadioGroup` component.
