+++ packages/web-ui/src/RadioGroup/index.ts
@@ -2,3 +2,5 @@ export { RadioGroup } from './RadioGroup';
 export type { RadioGroupProps } from './RadioGroup';
 export { RadioItem } from './RadioItem';
 export type { RadioItemProps } from './RadioItem';
+export { RadioTile } from './RadioTile';
+export type { RadioTileProps } from './RadioTile';
