+++ packages/web-ui/src/TextField/Textfield.stories.tsx
@@ -1,6 +1,4 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { colorsCommon } from '@utilitywarehouse/colour-system';
-import { Box } from '../Box';
 import { TextField } from './TextField';
 
 const meta: Meta<typeof TextField> = {
