+++ packages/web-ui/src/ToggleButton/ToggleButtonGroup.stories.tsx
@@ -1,7 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { colorsCommon } from '@utilitywarehouse/colour-system';
 import { useState } from 'react';
-import { Box } from '../Box';
 import { ToggleButton } from './ToggleButton';
 import { ToggleButtonGroup } from './ToggleButtonGroup';
